+++ custom_components/recalbox/api.py
@@ -89,6 +89,27 @@ class RecalboxAPI:
                 return False
 
 
+    async def is_kodi_running(self) -> bool:
+        kodi_url = f"http://{self.host}:{self.api_port_kodi}/jsonrpc"
+        payload = {
+            "jsonrpc": "2.0",
+            "method": "JSONRPC.Ping",
+            "id": 1
+        }
+        _LOGGER.debug(f"Ping Kodi : {kodi_url}")
+        connector = aiohttp.TCPConnector(family=socket.AF_INET) # Force la résolution en IPv4
+        async with aiohttp.ClientSession(connector=connector) as session:
+            try:
+                async with session.post(kodi_url, json=payload, timeout=5) as response:
+                    if response.status == 200:
+                        return True
+            except:
+                _LOGGER.info(f"Failed to ping Kodi via JSON RPC on {kodi_url}")
+                return False
+
+
+    # On va interroger Recalbox pour connaitre le status.
+    # S'il répond pas, on va quand même regarder si Kodi est lancé au démarrage
     async def get_current_status(self):
         url = f"http://{self.host}:{self.api_port_gamesmanager}/api/status"
         _LOGGER.debug(f"API GET current Recalbox status {url}")
@@ -146,8 +167,25 @@ class RecalboxAPI:
                             "status": "ON"
                         }
             except:
-                _LOGGER.error(f"Failed to get recalbox status on {url}")
-                raise
+                _LOGGER.error(f"Failed to get recalbox status on API {url}")
+                if (await self.is_kodi_running()) :
+                    _LOGGER.debug(f"Kodi seems to be running ! Simulating JSON data for Recalbox HA status")
+                    return {
+                        "game": None,
+                        "console": "Kodi",
+                        "rom": None,
+                        "genre": None,
+                        "genreId": None,
+                        "imagePath": None,
+                        "recalboxIpAddress": None,
+                        "recalboxVersion": None,
+                        "hardware": None,
+                        "scriptVersion": None,
+                        "status": "ON"
+                    }
+                else:
+                    _LOGGER.error(f"Kodi is not reachable neither")
+                    raise
 
     async def ping(self) -> bool:
         """Exécute un ping système vers l'hôte."""