+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Enable (re)configuration of the Kodi port
 - [#9](https://github.com/ooree23/RecalboxHomeAssistant/issues/9) If Recalbox didn't send any message, but suddenly responds to Ping : ask for current status to update the recalbox card
+- [#10](https://github.com/ooree23/RecalboxHomeAssistant/issues/10) If, when we try to pull status from Recalbox, the API call is KO, we try to Ping Kodi, to check if Recalbox started with Kodi and is running
 
 
 ## v1.5.1 - 04/02/2026 - Can now quit Kodi from HA